genesis-agent/agent_backend/live-agent-server-v2/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-firestore[async]
httpx